+++ backend/tests/integration/test_deliverable_template_http_api/test_create/test_dag.py
@@ -1,15 +1,15 @@
-"""DeliverableTemplate HTTP API 結合テスト — CREATE 系 (TC-IT-DTH-001〜006 / 018 / 019 / 021).
+"""DeliverableTemplate HTTP API 結合テスト — CREATE DAG 検証系 (TC-IT-DTH-002〜005 / 022).
 
 Covers:
-  TC-IT-DTH-001  POST /api/deliverable-templates → 201 全フィールド確認 (REQ-DT-HTTP-001)
   TC-IT-DTH-002  POST composition ref 不在 → 422 ref_not_found (MSG-DT-HTTP-002)
   TC-IT-DTH-003  POST 自己参照循環 → 422 validation_error (MSG-DT-002 domain invariant)
-  TC-IT-DTH-004  POST 推移的循環 A→B→A → 422 composition_cycle / transitive_cycle (MSG-DT-HTTP-003a)
+  TC-IT-DTH-004  POST 推移的循環 A→B→A → 422 composition_cycle / transitive_cycle
   TC-IT-DTH-005  POST DAG 深度ガード 11 → 422 composition_cycle / depth_limit (MSG-DT-HTTP-003b)
-  TC-IT-DTH-006  POST name 空文字 → 422 Pydantic validation
-  TC-IT-DTH-018  POST JSON_SCHEMA type → schema が dict で返却 (§確定I)
-  TC-IT-DTH-019  POST AcceptanceCriterion id 省略 → UUID 自動生成 (§確定H)
-  TC-IT-DTH-021  POST acceptance_criteria id 重複 → 422 domain invariant (§確定H)
+  TC-IT-DTH-022  POST 合法な菱形 DAG（A→B, A→C, B→C）→ 201 (DFS + 経路スタック正常系)
+
+セキュリティ注記:
+  composition_cycle エラーレスポンスは cycle_path を含まない（Tabriz 指摘 / OWASP A05 対応）。
+  cycle_path は内部 UUID 列挙を引き起こす可能性があるため最小情報開示原則に基づき除外済み。
 
 Issue: #122
 """
@@ -17,85 +17,16 @@ Issue: #122
 from __future__ import annotations
 
 import uuid
-from typing import Any
 
 import pytest
 
 from tests.integration.test_deliverable_template_http_api.conftest import DtTestCtx
+from tests.integration.test_deliverable_template_http_api.test_create.conftest import (
+    _MINIMAL_MARKDOWN_BODY,
+    _create_template,
+)
 
 pytestmark = pytest.mark.asyncio
-
-# ---------------------------------------------------------------------------
-# ヘルパ
-# ---------------------------------------------------------------------------
-
-_MINIMAL_MARKDOWN_BODY: dict[str, Any] = {
-    "name": "テストテンプレート",
-    "description": "説明",
-    "type": "MARKDOWN",
-    "schema": "## ガイドライン",
-    "version": {"major": 1, "minor": 0, "patch": 0},
-    "acceptance_criteria": [],
-    "composition": [],
-}
-
-
-async def _create_template(
-    ctx: DtTestCtx,
-    body: dict[str, Any] | None = None,
-    *,
-    assert_201: bool = True,
-) -> dict[str, Any]:
-    """POST /api/deliverable-templates してパース済み JSON ボディを返す。"""
-    payload = body if body is not None else _MINIMAL_MARKDOWN_BODY
-    resp = await ctx.client.post("/api/deliverable-templates", json=payload)
-    if assert_201:
-        assert resp.status_code == 201, f"template creation failed: {resp.text}"
-    return resp.json()  # type: ignore[return-value]
-
-
-# ---------------------------------------------------------------------------
-# TC-IT-DTH-001: POST 正常系
-# ---------------------------------------------------------------------------
-class TestCreateTemplate:
-    """TC-IT-DTH-001: POST /api/deliverable-templates → 201 (REQ-DT-HTTP-001)。"""
-
-    async def test_create_returns_201(self, dt_ctx: DtTestCtx) -> None:
-        resp = await dt_ctx.client.post("/api/deliverable-templates", json=_MINIMAL_MARKDOWN_BODY)
-        assert resp.status_code == 201
-
-    async def test_create_response_id_is_uuid_string(self, dt_ctx: DtTestCtx) -> None:
-        body = await _create_template(dt_ctx)
-        # UUID 形式であることを uuid.UUID でパース確認
-        parsed = uuid.UUID(body["id"])
-        assert isinstance(parsed, uuid.UUID)
-
-    async def test_create_response_name_matches(self, dt_ctx: DtTestCtx) -> None:
-        body = await _create_template(dt_ctx)
-        assert body["name"] == _MINIMAL_MARKDOWN_BODY["name"]
-
-    async def test_create_response_type_matches(self, dt_ctx: DtTestCtx) -> None:
-        body = await _create_template(dt_ctx)
-        assert body["type"] == "MARKDOWN"
-
-    async def test_create_response_version_matches(self, dt_ctx: DtTestCtx) -> None:
-        body = await _create_template(dt_ctx)
-        assert body["version"] == {"major": 1, "minor": 0, "patch": 0}
-
-    async def test_create_response_acceptance_criteria_empty(self, dt_ctx: DtTestCtx) -> None:
-        body = await _create_template(dt_ctx)
-        assert body["acceptance_criteria"] == []
-
-    async def test_create_response_composition_empty(self, dt_ctx: DtTestCtx) -> None:
-        body = await _create_template(dt_ctx)
-        assert body["composition"] == []
-
-    async def test_create_round_trip_get_returns_same_id(self, dt_ctx: DtTestCtx) -> None:
-        """ラウンドトリップ: POST → GET で同一 id が確認できる。"""
-        created = await _create_template(dt_ctx)
-        get_resp = await dt_ctx.client.get(f"/api/deliverable-templates/{created['id']}")
-        assert get_resp.status_code == 200
-        assert get_resp.json()["id"] == created["id"]
 
 
 # ---------------------------------------------------------------------------
@@ -140,7 +71,7 @@ class TestCreateSelfReference:
     """
 
     async def test_create_with_self_reference_returns_422(self, dt_ctx: DtTestCtx) -> None:
-        """DB に先に INSERT してから自己参照 POST を試みる。"""
+        """DB に先に INSERT してから自己参照 PUT を試みる。"""
         # まず template を 1 件作成して id を取得
         created = await _create_template(dt_ctx)
         template_id = created["id"]
@@ -181,7 +112,7 @@ class TestCreateSelfReference:
 # TC-IT-DTH-004: POST — 推移的循環 A→B→A → 422 transitive_cycle
 # ---------------------------------------------------------------------------
 class TestCreateTransitiveCycle:
-    """TC-IT-DTH-004: 推移的循環参照 → 422 composition_cycle / transitive_cycle (MSG-DT-HTTP-003a)。
+    """TC-IT-DTH-004: 推移的循環参照 → 422 composition_cycle / transitive_cycle (MSG-DT-HTTP-003a).
 
     セットアップ:
       1. B を作成（composition なし）
@@ -282,7 +213,11 @@ class TestCreateTransitiveCycle:
         )
         assert resp.json()["error"]["detail"]["reason"] == "transitive_cycle"
 
-    async def test_transitive_cycle_cycle_path_contains_both_uuids(self, dt_ctx: DtTestCtx) -> None:
+    async def test_transitive_cycle_no_cycle_path_in_response(self, dt_ctx: DtTestCtx) -> None:
+        """cycle_path はセキュリティ上レスポンスに含まれないこと（OWASP A05 / Tabriz 指摘対応）。
+
+        旧実装では cycle_path に内部 UUID を含んでいたが、最小情報開示原則に基づき除外された。
+        """
         b = await _create_template(dt_ctx, {**_MINIMAL_MARKDOWN_BODY, "name": "template-B4"})
         a = await _create_template(
             dt_ctx,
@@ -310,15 +245,15 @@ class TestCreateTransitiveCycle:
                 ],
             },
         )
-        cycle_path = resp.json()["error"]["detail"]["cycle_path"]
-        assert a["id"] in cycle_path or b["id"] in cycle_path
+        # cycle_path はレスポンス detail に含まれないこと（最小情報開示原則）
+        assert "cycle_path" not in resp.json()["error"]["detail"]
 
 
 # ---------------------------------------------------------------------------
 # TC-IT-DTH-005: POST — DAG 深度ガード → 422 depth_limit (§確定D)
 # ---------------------------------------------------------------------------
 class TestCreateDagDepthLimit:
-    """TC-IT-DTH-005: composition チェーン深度 11 → 422 depth_limit (MSG-DT-HTTP-003b)。
+    """TC-IT-DTH-005: composition チェーン深度 11 → 422 depth_limit (MSG-DT-HTTP-003b).
 
     T1(leaf) → T2 → ... → T11 の 11 段チェーンを構築し、
     T12 が T11 を参照すると depth 11 > 10 で depth_limit。
@@ -423,7 +358,10 @@ class TestCreateDagDepthLimit:
         )
         assert resp.json()["error"]["detail"]["reason"] == "depth_limit"
 
-    async def test_create_dag_depth_limit_cycle_path_is_empty(self, dt_ctx: DtTestCtx) -> None:
+    async def test_create_dag_depth_limit_no_cycle_path_in_response(
+        self, dt_ctx: DtTestCtx
+    ) -> None:
+        """depth_limit レスポンスにも cycle_path は含まれないこと（OWASP A05）。"""
         prev = await _create_template(dt_ctx, {**_MINIMAL_MARKDOWN_BODY, "name": "de-chain-T1"})
         for i in range(2, 12):
             prev = await _create_template(
@@ -452,98 +390,141 @@ class TestCreateDagDepthLimit:
                 ],
             },
         )
-        assert resp.json()["error"]["detail"]["cycle_path"] == []
+        # cycle_path はレスポンス detail に含まれないこと（最小情報開示原則）
+        assert "cycle_path" not in resp.json()["error"]["detail"]
 
 
 # ---------------------------------------------------------------------------
-# TC-IT-DTH-006: POST — name 空文字 → 422
+# TC-IT-DTH-022: POST — 合法な菱形 DAG（A→B, A→C, B→C）→ 201 (§確定D)
 # ---------------------------------------------------------------------------
-class TestCreateInvalidName:
-    """TC-IT-DTH-006: name 空文字 → 422 Pydantic バリデーション。"""
+class TestCreateDiamondDag:
+    """TC-IT-DTH-022: 合法な菱形 DAG（diamond DAG）が 201 で正常登録されること (§確定D).
 
-    async def test_create_with_empty_name_returns_422(self, dt_ctx: DtTestCtx) -> None:
-        body = {**_MINIMAL_MARKDOWN_BODY, "name": ""}
-        resp = await dt_ctx.client.post("/api/deliverable-templates", json=body)
-        assert resp.status_code == 422
+    セットアップ:
+      1. C を作成（composition なし、leaf node）
+      2. B を作成（composition = [ref_to_C]）  → B→C (正常)
+      3. A を作成（composition = [ref_to_B, ref_to_C]）→ A→B, A→C, B→C (菱形DAG)
 
+    旧 BFS 実装では C が B 経由と A 直接経由の 2 経路で visited に 2 回現れ、
+    誤って循環と判定していた（ヘルスバーグ指摘 Critical Defect #1）。
+    新 DFS + ancestors（経路スタック）実装では ancestors が現在の DFS 経路のみを
+    保持するため、C の再訪問は祖先集合に入らず循環として誤検出されない。
+    """
 
-# ---------------------------------------------------------------------------
-# TC-IT-DTH-018: POST — JSON_SCHEMA type → schema が dict で返却 (§確定I)
-# ---------------------------------------------------------------------------
-class TestCreateJsonSchemaType:
-    """TC-IT-DTH-018: type=JSON_SCHEMA → schema レスポンスが dict (§確定I)。"""
-
-    async def test_create_json_schema_type_returns_dict_schema(self, dt_ctx: DtTestCtx) -> None:
-        """JSON_SCHEMA type で dict schema を POST → レスポンスの schema が dict。"""
-        from bakufu.domain.deliverable_template.deliverable_template import DeliverableTemplate
-
-        from tests.factories.deliverable_template import ValidStubValidator
-
-        # テスト用 validator を注入（Fail Secure 回避）
-        original_validator = DeliverableTemplate._validator  # pyright: ignore[reportPrivateUsage]
-        try:
-            DeliverableTemplate._validator = ValidStubValidator()  # pyright: ignore[reportPrivateUsage]
-            body = {
+    async def test_diamond_dag_returns_201(self, dt_ctx: DtTestCtx) -> None:
+        """A→B, A→C, B→C の菱形 DAG が 201 で登録できること（循環エラーにならないこと）。"""
+        # 1. C 作成（leaf）
+        c = await _create_template(dt_ctx, {**_MINIMAL_MARKDOWN_BODY, "name": "diamond-C"})
+        # 2. B 作成（B→C）
+        b = await _create_template(
+            dt_ctx,
+            {
                 **_MINIMAL_MARKDOWN_BODY,
-                "type": "JSON_SCHEMA",
-                "schema": {"$schema": "http://json-schema.org/draft-07/schema#", "type": "object"},
-            }
-            resp = await dt_ctx.client.post("/api/deliverable-templates", json=body)
-            assert resp.status_code == 201
-            assert isinstance(resp.json()["schema"], dict)
-        finally:
-            DeliverableTemplate._validator = original_validator  # pyright: ignore[reportPrivateUsage]
+                "name": "diamond-B",
+                "composition": [
+                    {
+                        "template_id": c["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    }
+                ],
+            },
+        )
+        # 3. A 作成（A→B, A→C → 菱形DAG）
+        resp = await dt_ctx.client.post(
+            "/api/deliverable-templates",
+            json={
+                **_MINIMAL_MARKDOWN_BODY,
+                "name": "diamond-A",
+                "composition": [
+                    {
+                        "template_id": b["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    },
+                    {
+                        "template_id": c["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    },
+                ],
+            },
+        )
+        assert resp.status_code == 201, (
+            f"diamond DAG は合法な DAG であり composition_cycle になってはならない: {resp.text}"
+        )
 
+    async def test_diamond_dag_does_not_return_composition_cycle(self, dt_ctx: DtTestCtx) -> None:
+        """A→B, A→C, B→C → composition_cycle エラーにならないこと。"""
+        c = await _create_template(dt_ctx, {**_MINIMAL_MARKDOWN_BODY, "name": "diamond-C2"})
+        b = await _create_template(
+            dt_ctx,
+            {
+                **_MINIMAL_MARKDOWN_BODY,
+                "name": "diamond-B2",
+                "composition": [
+                    {
+                        "template_id": c["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    }
+                ],
+            },
+        )
+        resp = await dt_ctx.client.post(
+            "/api/deliverable-templates",
+            json={
+                **_MINIMAL_MARKDOWN_BODY,
+                "name": "diamond-A2",
+                "composition": [
+                    {
+                        "template_id": b["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    },
+                    {
+                        "template_id": c["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    },
+                ],
+            },
+        )
+        # 旧 BFS 実装では "composition_cycle" になっていたが DFS 実装では正常 → 検証
+        assert resp.json().get("error", {}).get("code") != "composition_cycle", (
+            f"diamond DAG が誤って composition_cycle と判定された: {resp.text}"
+        )
 
-# ---------------------------------------------------------------------------
-# TC-IT-DTH-019: POST — AcceptanceCriterion id 省略 → UUID 自動生成 (§確定H)
-# ---------------------------------------------------------------------------
-class TestCreateWithOmittedAcId:
-    """TC-IT-DTH-019: acceptance_criteria.id 省略 → UUID 自動生成 (§確定H)。"""
-
-    async def test_create_with_omitted_ac_id_generates_uuid(self, dt_ctx: DtTestCtx) -> None:
-        body = {
-            **_MINIMAL_MARKDOWN_BODY,
-            "acceptance_criteria": [
-                {"description": "条件1", "required": True},
-            ],
-        }
-        resp = await dt_ctx.client.post("/api/deliverable-templates", json=body)
+    async def test_diamond_dag_response_has_two_composition_refs(self, dt_ctx: DtTestCtx) -> None:
+        """A の composition に B と C の 2 つの ref が含まれること。"""
+        c = await _create_template(dt_ctx, {**_MINIMAL_MARKDOWN_BODY, "name": "diamond-C3"})
+        b = await _create_template(
+            dt_ctx,
+            {
+                **_MINIMAL_MARKDOWN_BODY,
+                "name": "diamond-B3",
+                "composition": [
+                    {
+                        "template_id": c["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    }
+                ],
+            },
+        )
+        resp = await dt_ctx.client.post(
+            "/api/deliverable-templates",
+            json={
+                **_MINIMAL_MARKDOWN_BODY,
+                "name": "diamond-A3",
+                "composition": [
+                    {
+                        "template_id": b["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    },
+                    {
+                        "template_id": c["id"],
+                        "minimum_version": {"major": 1, "minor": 0, "patch": 0},
+                    },
+                ],
+            },
+        )
         assert resp.status_code == 201
-        ac_id = resp.json()["acceptance_criteria"][0]["id"]
-        # UUID 形式であることを確認
-        parsed = uuid.UUID(ac_id)
-        assert isinstance(parsed, uuid.UUID)
-
-
-# ---------------------------------------------------------------------------
-# TC-IT-DTH-021: POST — acceptance_criteria id 重複 → 422 (§確定H)
-# ---------------------------------------------------------------------------
-class TestCreateWithDuplicateAcId:
-    """TC-IT-DTH-021: acceptance_criteria 内 id 重複 → 422 domain invariant (§確定H)。"""
-
-    async def test_create_with_duplicate_ac_id_returns_422(self, dt_ctx: DtTestCtx) -> None:
-        dup_id = str(uuid.uuid4())
-        body = {
-            **_MINIMAL_MARKDOWN_BODY,
-            "acceptance_criteria": [
-                {"id": dup_id, "description": "条件1", "required": True},
-                {"id": dup_id, "description": "条件2", "required": True},
-            ],
-        }
-        resp = await dt_ctx.client.post("/api/deliverable-templates", json=body)
-        assert resp.status_code == 422
-
-    async def test_create_with_duplicate_ac_id_code_is_validation_error(
-        self, dt_ctx: DtTestCtx
-    ) -> None:
-        dup_id = str(uuid.uuid4())
-        body = {
-            **_MINIMAL_MARKDOWN_BODY,
-            "acceptance_criteria": [
-                {"id": dup_id, "description": "条件1", "required": True},
-                {"id": dup_id, "description": "条件2", "required": True},
-            ],
-        }
-        resp = await dt_ctx.client.post("/api/deliverable-templates", json=body)
-        assert resp.json()["error"]["code"] == "validation_error"
+        composition = resp.json()["composition"]
+        assert len(composition) == 2
+        template_ids = {ref["template_id"] for ref in composition}
+        assert b["id"] in template_ids
+        assert c["id"] in template_ids
